requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
python-dotenv>=1.0.1
openai>=1.40.0
streamlit
pandas
numpy
python-dotenv
transformers
torch
streamlit>=1.36.0
pandas>=2.0.0
transformers>=4.41.0
sentencepiece>=0.1.99
sentence-transformers>=3.0.1
scikit-learn>=1.4.0
torch
google-generativeai>=0.8.0

